# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "yunq API",
    "description": "Static analysis platform: REST API, OAuth, signed webhooks and operational metrics.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.1"
  },
  "paths": {
...
    "/api/projects/{key}/components/tree": {
      "get": {
        "summary": "A project's components (currently: files only — see module docs) with\ntheir latest measures, as of the most recent analysis. Flat navigation,\nminus directory nesting.",
        "operationId": "component_tree",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Project key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branch",
            "in": "query",
            "description": "Branch the tree is scoped to (default \"main\").",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort field: `name` or a measure key present on the components, e.g.\n`issue_total` (default \"name\").",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "description": "Sort direction: `asc` or `desc` (default \"asc\").",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Case-sensitive substring filter on file path.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The project's known files with their latest measures",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComponentTreeDto"
                }
              }
            }
          },
          "404": {
            "description": "No analysis exists yet for this project/branch — run a scan first"
          },
          "502": {
            "description": "Storage backend failure"
          }
        }
      }
    },
...
    "/api/projects/{key}/measures/history": {
      "get": {
        "summary": "Time series of a project's (or one of its files') measures across its\nanalyses.",
        "operationId": "measure_history",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Project key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branch",
            "in": "query",
            "description": "Branch the history is scoped to (default \"main\").",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "component",
            "in": "query",
            "description": "File path to scope the history to one file's measures instead of\nthe project as a whole.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "metrics",
            "in": "query",
            "description": "Comma-separated metric keys, e.g. `coverage,issue_total` (default:\nevery metric persisted for that analysis).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Inclusive lower bound, ISO-8601 (e.g. `2024-01-01T00:00:00Z`).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Inclusive upper bound, ISO-8601.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Measure time series, oldest analysis first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasureHistoryDto"
                }
              }
            }
          },
          "502": {
            "description": "Storage backend failure"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComponentDto": {
        "type": "object",
        "required": [
          "path",
          "measures"
        ],
        "properties": {
          "measures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MeasureValueDto"
            }
          },
          "path": {
            "type": "string"
          }
        }
      },
      "ComponentTreeDto": {
        "type": "object",
        "required": [
          "analysis_id",
          "total",
          "components",
          "tree"
        ],
        "properties": {
          "analysis_id": {
            "type": "integer",
            "format": "int64"
          },
          "components": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComponentDto"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "tree": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComponentTreeNodeDto"
            }
          }
        }
      },
      "ComponentTreeNodeDto": {
        "type": "object",
        "description": "One node of the nested `tree` view: a directory (`qualifier: \"DIR\"`,\n`children` populated, `measures` empty) or a file (`qualifier: \"FIL\"`,\n`measures` populated, no children). The DIR/FIL qualifier vocabulary is\nthe conventional one, so existing tooling recognizes the field.",
        "required": [
          "name",
          "path",
          "qualifier"
        ],
        "properties": {
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComponentTreeNodeDto"
            }
          },
          "measures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MeasureValueDto"
            }
          },
          "name": {
            "type": "string",
            "description": "Final path segment (directory or file name)."
          },
          "path": {
            "type": "string",
            "description": "Full path from the project root."
          },
          "qualifier": {
            "type": "string"
          }
        }
      },
...
      "MeasureHistoryDto": {
        "type": "object",
        "required": [
          "component",
          "history"
        ],
        "properties": {
          "component": {
            "type": "string",
            "description": "The project key, or `{project key}:{file path}` when `component`\nwas given in the request."
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MeasureHistoryPointDto"
            }
          }
        }
      },
      "MeasureHistoryPointDto": {
        "type": "object",
        "required": [
          "analysis_id",
          "date",
          "measures"
        ],
        "properties": {
          "analysis_id": {
            "type": "integer",
            "format": "int64"
          },
          "date": {
            "type": "string",
            "description": "ISO-8601 timestamp of the analysis this point belongs to."
          },
          "measures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MeasureValueDto"
            }
          }
        }
      },
      "MeasureValueDto": {
        "type": "object",
        "required": [
          "metric",
          "value"
        ],
        "properties": {
          "metric": {
            "type": "string"
          },
          "value": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}